tidy up and json made
