+++ src/Bookings.js
@@ -6,7 +6,6 @@ const Bookings = () => {
   const search = searchVal => {
     console.info("TO DO!", searchVal);
   };
-
   return (
     <div className="App-content">
       <div className="container">
