+++ src/Heading.js
@@ -1,0 +1,11 @@
+const Heading = () => (
+  <header className="App-header">
+    <img
+      className="App-logo"
+      src="https://img.freepik.com/premium-vector/hotel-logo-simple-illustration_434503-736.jpg?w=2000"
+    />
+    <span className="App-title">CYF Hotel</span>
+  </header>
+);
+
+export default Heading;
